+++ src/main/resources/swagger/api-docs.yaml
@@ -20,7 +20,7 @@ security:
       - email
       - profile
 paths:
-  /api/admin/user/{userId}:
+  /api/admin/v1/user/{userId}:
     get:
       tags:
         - Admin
@@ -40,7 +40,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/UserStatus'
-  /api/admin/user/email/{email}:
+  /api/admin/v1/user/email/{email}:
     get:
       tags:
         - Admin
@@ -60,7 +60,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/UserStatus'
-  /api/admin/user/{userId}/disable:
+  /api/admin/v1/user/{userId}/disable:
     put:
       tags:
         - Admin
@@ -80,7 +80,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/UserStatus'
-  /api/admin/user/{userId}/enable:
+  /api/admin/v1/user/{userId}/enable:
     put:
       tags:
         - Admin
@@ -100,7 +100,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/UserStatus'
-  /api/admin/user/{userId}/petServiceAccount:
+  /api/admin/v1/user/{userId}/petServiceAccount:
     delete:
       tags:
         - Admin
@@ -126,7 +126,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorReport'
-  /api/admin/user/{userId}/petServiceAccount/{project}:
+  /api/admin/v1/user/{userId}/petServiceAccount/{project}:
     delete:
       tags:
         - Admin
@@ -158,7 +158,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorReport'
-  /api/resourceTypeAdmin/v1/resources/{resourceTypeName}/{resourceId}/policies:
+  /api/admin/v1/resources/{resourceTypeName}/{resourceId}/policies:
     get:
       tags:
         - Admin
@@ -199,7 +199,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorReport'
-  /api/resourceTypeAdmin/v1/resources/{resourceTypeName}/{resourceId}/policies/{policyName}/memberEmails/{userEmail}:
+  /api/admin/v1/resources/{resourceTypeName}/{resourceId}/policies/{policyName}/memberEmails/{userEmail}:
     put:
       tags:
         - Admin
@@ -300,7 +300,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorReport'
-  /api/resourceTypeAdmin/v1/resourceTypes/{resourceTypeName}/policies:
+  /api/admin/v1/resourceTypes/{resourceTypeName}/policies:
     get:
       tags:
         - Admin
@@ -334,7 +334,7 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorReport'
-  /api/resourceTypeAdmin/v1/resourceTypes/{resourceTypeName}/policies/{policyName}:
+  /api/admin/v1/resourceTypes/{resourceTypeName}/policies/{policyName}:
     put:
       tags:
         - Admin
