+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/api/SamRoutes.scala
@@ -48,7 +48,7 @@ abstract class SamRoutes(
     with TermsOfServiceRoutes
     with ExtensionRoutes
     with ManagedGroupRoutes
-    with AdminResourceRoutes {
+    with AdminRoutes {
 
   def route: server.Route = (logRequestResult & handleExceptions(myExceptionHandler)) {
     oidcConfig.swaggerRoutes("swagger/api-docs.yaml") ~
@@ -63,7 +63,7 @@ abstract class SamRoutes(
             withActiveUser(samRequestContext) { samUser =>
               val samRequestContextWithUser = samRequestContext.copy(samUser = Option(samUser))
               resourceRoutes(samUser, samRequestContextWithUser) ~
-                adminResourceRoutes(samUser, samRequestContextWithUser) ~
+                adminRoutes(samUser, samRequestContextWithUser) ~
                 adminUserRoutes(samUser, samRequestContextWithUser) ~
                 extensionRoutes(samUser, samRequestContextWithUser) ~
                 groupRoutes(samUser, samRequestContextWithUser) ~
