+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/api/UserRoutes.scala
@@ -6,7 +6,6 @@ import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.{Directive0, ExceptionHandler}
-import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 import org.broadinstitute.dsde.workbench.model._
 import org.broadinstitute.dsde.workbench.sam.model.SamJsonSupport._
 import org.broadinstitute.dsde.workbench.sam.model.SamUser
@@ -146,86 +145,6 @@ trait UserRoutes extends SamUserDirectives with SamRequestContextDirectives {
                             StatusCodes.OK -> Option(status)
                           }
                           .getOrElse(StatusCodes.NotFound -> None)
-                      }
-                    }
-                  }
-                }
-            }
-        }
-      }
-    }
-
-  def adminUserRoutes(samUser: SamUser, samRequestContext: SamRequestContext): server.Route =
-    pathPrefix("admin") {
-      withWorkbenchAdmin(samUser) {
-        pathPrefix("user") {
-          path("email" / Segment) { email =>
-            complete {
-              userService.getUserStatusFromEmail(WorkbenchEmail(email), samRequestContext).map { statusOption =>
-                statusOption
-                  .map { status =>
-                    StatusCodes.OK -> Option(status)
-                  }
-                  .getOrElse(StatusCodes.NotFound -> None)
-              }
-            }
-          } ~
-            pathPrefix(Segment) { userId =>
-              pathEnd {
-                delete {
-                  complete {
-                    userService.deleteUser(WorkbenchUserId(userId), samRequestContext).map(_ => StatusCodes.OK)
-                  }
-                } ~
-                  get {
-                    complete {
-                      userService.getUserStatus(WorkbenchUserId(userId), samRequestContext = samRequestContext).map { statusOption =>
-                        statusOption
-                          .map { status =>
-                            StatusCodes.OK -> Option(status)
-                          }
-                          .getOrElse(StatusCodes.NotFound -> None)
-                      }
-                    }
-                  }
-              } ~
-                pathPrefix("enable") {
-                  pathEndOrSingleSlash {
-                    put {
-                      complete {
-                        userService.enableUser(WorkbenchUserId(userId), samRequestContext).map { statusOption =>
-                          statusOption
-                            .map { status =>
-                              StatusCodes.OK -> Option(status)
-                            }
-                            .getOrElse(StatusCodes.NotFound -> None)
-                        }
-                      }
-                    }
-                  }
-                } ~
-                pathPrefix("disable") {
-                  pathEndOrSingleSlash {
-                    put {
-                      complete {
-                        userService.disableUser(WorkbenchUserId(userId), samRequestContext).map { statusOption =>
-                          statusOption
-                            .map { status =>
-                              StatusCodes.OK -> Option(status)
-                            }
-                            .getOrElse(StatusCodes.NotFound -> None)
-                        }
-                      }
-                    }
-                  }
-                } ~
-                pathPrefix("petServiceAccount") {
-                  path(Segment) { project =>
-                    delete {
-                      complete {
-                        cloudExtensions
-                          .deleteUserPetServiceAccount(WorkbenchUserId(userId), GoogleProject(project), samRequestContext)
-                          .map(_ => StatusCodes.NoContent)
                       }
                     }
                   }
