+++ src/test/scala/org/broadinstitute/dsde/workbench/sam/api/UserRoutesV1Spec.scala
@@ -16,7 +16,7 @@ import org.broadinstitute.dsde.workbench.sam.service.{NoExtensions, StatusServic
 /**
   * Created by dvoet on 6/7/17.
   */
-class UserRoutesV1Spec extends UserRoutesSpecHelper{
+class UserRoutesV1Spec extends UserRoutesSpecHelper {
 
   def withSARoutes[T](testCode: (TestSamRoutes, TestSamRoutes) => T): T = {
     val directoryDAO = new MockDirectoryDAO()
@@ -124,103 +124,6 @@ class UserRoutesV1Spec extends UserRoutesSpecHelper{
     Get("/register/user/v1?userDetailsOnly=true") ~> routes.route ~> check {
       status shouldEqual StatusCodes.OK
       responseAs[UserStatus] shouldEqual UserStatus(UserStatusDetails(user.id, user.email), Map.empty)
-    }
-  }
-
-  "GET /admin/user/{userSubjectId}" should "get the user status of a user (as an admin)" in {
-    val (user, getRoutes) = setUpAdminTest()
-    Get(s"/api/admin/user/${user.id}") ~> getRoutes.route ~> check {
-      status shouldEqual StatusCodes.OK
-      responseAs[UserStatus] shouldEqual UserStatus(UserStatusDetails(user.id, user.email), Map("ldap" -> true, "allUsersGroup" -> true, "google" -> true))
-    }
-  }
-
-  it should "not allow a non-admin with userId to get the status of another user" in withAdminRoutes { (samRoutes, _) =>
-    Get(s"/api/admin/user/$defaultUserId") ~> samRoutes.route ~> check {
-      status shouldEqual StatusCodes.Forbidden
-    }
-  }
-
-  "GET /admin/user/email/{email}" should "get the user status of a user by email (as an admin)" in {
-    val (user, getRoutes) = setUpAdminTest()
-
-    Get(s"/api/admin/user/email/${user.email}") ~> getRoutes.route ~> check {
-      status shouldEqual StatusCodes.OK
-      responseAs[UserStatus] shouldEqual UserStatus(UserStatusDetails(user.id, user.email), Map("ldap" -> true, "allUsersGroup" -> true, "google" -> true))
-    }
-  }
-
-  it should "return 404 for an unknown user by email (as an admin)" in withAdminRoutes { (samRoutes, adminRoutes) =>
-    Get(s"/api/admin/user/email/XXX${defaultUserEmail}XXX") ~> adminRoutes.route ~> check {
-      status shouldEqual StatusCodes.NotFound
-    }
-  }
-
-  it should "return 404 for an group's email (as an admin)" in withAdminRoutes { (samRoutes, adminRoutes) =>
-    Get(s"/api/admin/user/email/fc-admins@dev.test.firecloud.org") ~> adminRoutes.route ~> check {
-      status shouldEqual StatusCodes.NotFound
-    }
-  }
-
-  it should "not allow a non-admin with user email to get the status of another user" in withAdminRoutes { (samRoutes, _) =>
-    Get(s"/api/admin/user/email/$defaultUserEmail") ~> samRoutes.route ~> check {
-      status shouldEqual StatusCodes.Forbidden
-    }
-  }
-
-  "PUT /admin/user/{userSubjectId}/(re|dis)able" should "disable and then re-enable a user (as an admin)" in {
-    val (user, adminRoutes) = setUpAdminTest()
-
-    Put(s"/api/admin/user/${user.id}/disable") ~> adminRoutes.route ~> check {
-      status shouldEqual StatusCodes.OK
-      responseAs[UserStatus] shouldEqual UserStatus(UserStatusDetails(user.id, user.email), Map("ldap" -> false, "allUsersGroup" -> true, "google" -> true))
-    }
-
-    Put(s"/api/admin/user/${user.id}/enable") ~> adminRoutes.route ~> check {
-      status shouldEqual StatusCodes.OK
-      responseAs[UserStatus] shouldEqual UserStatus(UserStatusDetails(user.id, user.email), Map("ldap" -> true, "allUsersGroup" -> true, "google" -> true))
-    }
-  }
-
-  it should "not allow a non-admin to enable or disable a user" in withAdminRoutes { (samRoutes, _) =>
-    Put(s"/api/admin/user/$defaultUserId/disable") ~> samRoutes.route ~> check {
-      status shouldEqual StatusCodes.Forbidden
-    }
-
-    Put(s"/api/admin/user/$defaultUserId/enable") ~> samRoutes.route ~> check {
-      status shouldEqual StatusCodes.Forbidden
-    }
-  }
-
-  "DELETE /admin/user/{userSubjectId}" should "delete a user (as an admin)" in {
-    val (user, adminRoutes) = setUpAdminTest()
-
-    Delete(s"/api/admin/user/${user.id}") ~> adminRoutes.route ~> check {
-      status shouldEqual StatusCodes.OK
-    }
-
-    Get(s"/api/admin/user/${user.id}")~> adminRoutes.route ~> check {
-      status shouldEqual StatusCodes.NotFound
-    }
-  }
-
-  it should "not allow a non-admin to delete a user" in withAdminRoutes { (samRoutes, _) =>
-    Delete(s"/api/admin/user/$defaultUserId") ~> samRoutes.route ~> check {
-      status shouldEqual StatusCodes.Forbidden
-    }
-  }
-
-  "DELETE /admin/user/{userSubjectId}/petServiceAccount/{project}" should "delete a pet (as an admin)" in withAdminRoutes { (samRoutes, adminRoutes) =>
-    val (user, adminRoutes) = setUpAdminTest()
-
-    Delete(s"/api/admin/user/${user.id}/petServiceAccount/myproject") ~> adminRoutes.route ~> check {
-      status shouldEqual StatusCodes.NoContent
-    }
-  }
-
-  it should "not allow a non-admin to delete a pet" in withAdminRoutes { (samRoutes, _) =>
-    Delete(s"/api/admin/user/$defaultUserId/petServiceAccount/myproject") ~> samRoutes.route ~> check {
-      status shouldEqual StatusCodes.Forbidden
     }
   }
